factor HcInfo out of events
it doesn't belong there.

so we parameterize `Events` by a type var `HI`, and then instantiate
that as `HcInfo` in main.
